backend/requirements.txt
# This file is auto-generated by uv
# Do not edit manually
django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
djangorestframework-simplejwt>=5.3.0
google-generativeai>=0.3.0
boto3>=1.34.0
gunicorn>=21.2.0
uvicorn[standard]>=0.25.0
celery>=5.3.0
kombu[sqs]>=5.3.0
django-celery-results>=2.5.0
django-celery-beat>=2.5.0
django-redis>=5.4.0
redis>=5.0.0
hiredis>=2.2.0
pycurl>=7.45.0
requests>=2.31.0
pyyaml>=6.0.0
# DynamoDB support
boto3>=1.34.0
botocore>=1.34.0
